abstracted plant details to plant component
